Refactor and organize
