+++ assets/src/components/cd/pipelines/utils/reduceGateStatuses.tsx
